back/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
PyMuPDF
pdfplumber
google-cloud-vision
python-multipart
